+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneAccountProtectionPolicy/MSFT_IntuneAccountProtectionPolicy.psm1
@@ -1,3 +1,5 @@
+Confirm-M365DSCModuleDependency -ModuleName 'MSFT_IntuneAccountProtectionPolicy'
+
 function Get-TargetResource
 {
     [CmdletBinding()]
@@ -809,3 +811,4 @@ function Export-TargetResource
 }
 
 Export-ModuleMember -Function *-TargetResource
+