+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneApplicationControlPolicyWindows10/MSFT_IntuneApplicationControlPolicyWindows10.psm1
@@ -1,3 +1,5 @@
+Confirm-M365DSCModuleDependency -ModuleName 'MSFT_IntuneApplicationControlPolicyWindows10'
+
 function Get-TargetResource
 {
     [CmdletBinding()]
@@ -587,3 +589,4 @@ function Export-TargetResource
 }
 
 Export-ModuleMember -Function *-TargetResource
+