+++ Modules/Microsoft365DSC/DSCResources/MSFT_SPOSharingSettings/MSFT_SPOSharingSettings.psm1
@@ -1,3 +1,5 @@
+Confirm-M365DSCModuleDependency -ModuleName 'MSFT_SPOSharingSettings'
+
 function Get-TargetResource
 {
     [CmdletBinding()]
@@ -872,3 +874,4 @@ function Export-TargetResource
 }
 
 Export-ModuleMember -Function *-TargetResource
+